refactor(sidebar): replace AppSidebar component with a new structure of sidebar components for better organization and maintainability delete(AppSidebar.tsx): remove old AppSidebar component to streamline the sidebar implementation add(AppSideBarCTA.tsx, AppSideBarData.tsx, AppSideBarHeader.tsx, AppSideBarProject.tsx): introduce new components to enhance sidebar functionality and clarity
