+++ client/src/components/sidebar/AppSideBarHeader.tsx
@@ -1,0 +1,5 @@
+import {FC} from 'react';
+
+export const AppSideBarHeader: FC = () => {
+    return <div>HEADER</div>;
+};
