+++ client/src/demos/demo.tsx
@@ -2,8 +2,8 @@ import {AppHeader} from '@/components/app/AppHeader';
 import {AppHelloWorld} from '@/components/app/AppHelloWorld';
 import {AppPlaceholder} from '@/components/app/AppPlaceholder';
 import {AppProviders} from '@/components/app/AppProviders';
-import {AppSidebar} from '@/components/app/AppSidebar';
 import {SidebarInset, SidebarProvider} from '@/components/shadcn/ui/sidebar';
+import {AppSidebar} from '@/components/sidebar/AppSidebar';
 import {FC, Suspense} from 'react';
 
 /**
